DELETE works

+++ app/item/controller.py
@@ -48,13 +48,15 @@ def index():
 @itemMod.route('/<string:itemName>', methods=['GET', 'PUT', 'DELETE'])
 def update(itemName):
     #get the Item.
-    item = get(itemName)
-    print(item)
-    if not item:
-        return jsonify('%r not found in database.' % itemName)
+
 
     # view a single Item
     if request.method == 'GET':
+        item = get(itemName)
+        print(item)
+        if not item:
+            return jsonify('%r not found in database.' % itemName)
+        item = a_dict(item)
         return jsonify({
             'item': item
         })
@@ -65,7 +67,12 @@ def update(itemName):
 
     # delete a single item
     if request.method == 'DELETE':
+        item = get(itemName)
+        print(item)
+        if not item:
+            return jsonify('%r not found in database.' % itemName)
         try:
+            #item = a_dict(item)
             db.session.delete(item)
             db.session.commit()
             return jsonify('%r deleted from the database.' % itemName)
@@ -78,10 +85,10 @@ def get(itemName):
     if itemName == "":
         return jsonify('error: itemName must not be a null string argument'), 404
     """
-    item = Item.query.filter(Item.name == itemName).all()
+    item = Item.query.filter_by(name=itemName).first()
     if not item:
         return False
-    item = item[0].to_dict()
+    #item = item[0].to_dict()
     return item
 
 def a_dict(item):
